+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
-import type { Metadata } from 'next';
-import { headers } from 'next/headers'; // headers 사용
-import HeadNav from '@/components/layout/HeadNav';
-import SideNav from '@/components/layout/SideNav';
-import '@/styles/globals.css';
+import type { Metadata } from 'next'
+import { headers } from 'next/headers' // headers 사용
+import HeadNav from '@/components/layout/HeadNav'
+import SideNav from '@/components/layout/SideNav'
+import '@/styles/globals.css'
 
 export const metadata: Metadata = {
   title: 'Duo-Dingco',
@@ -10,15 +10,15 @@ export const metadata: Metadata = {
   icons: {
     icon: '/dingco.png',
   },
-};
+}
 
 export default function RootLayout({
   children,
 }: Readonly<{
-  children: React.ReactNode;
+  children: React.ReactNode
 }>) {
-  const pathname = headers().get('x-pathname') || '/'; // 현재 경로 가져오기
-  const isAuthPage = pathname.startsWith('/auth'); // 인증 페이지 확인
+  const pathname = headers().get('x-pathname') || '/' // 현재 경로 가져오기
+  const isAuthPage = pathname.startsWith('/api/auth') // 인증 페이지 확인
 
   return (
     <html lang="en">
@@ -31,5 +31,5 @@ export default function RootLayout({
         </div>
       </body>
     </html>
-  );
+  )
 }
